Catch subprocess.CalledProcessError instead of OSError in `eden doctor`.
Summary:
Before I fixed an issue with `nuclide-connections` in D7833410, it could throw
an error. Although there was a try/except around the `check_output` call in
`eden doctor`, it caught `OSError` rather than `CalledProcessError`, which
seems like a mistake (or maybe an inadvertent evolution of the code).

While there, I also extended it to catch `ValueError` in case the stdout of the
subprocess is not valid JSON, in which case `json.loads()` will raise a
`ValueError`.

Reviewed By: chadaustin

Differential Revision: D7890571

fbshipit-source-id: 184f6f669e9d62a5fb04db29bcbab450defc226e