fsmonitor: unlink fsmonitor.state before writing dirstate
Summary:
This is an attempt to fix the wrong status issue. A formal fix would be
D7909172, which merges fsmonitor.state into treestate. See the comment in
`__init__.py` change for why status can be wrong.

Although it fixes the "status" path, I'm not sure whether other code paths
writing dirstates could be problematic or not. Anyway, D7909172 should be the
preferred final fix.

Reviewed By: r4-in

Differential Revision: D7916344

fbshipit-source-id: beab1c825b970bb47ffe03617a14eb6f203feafa