commitcloud: prevent race between creating and syncing obsmarkers
Summary:
In order to prevent newly added obsmarkers from being lost while we are syncing
them, we separate out the obsmarkers that we are in the process of syncing into
a separate `syncing` file.  New obsmarkers that are created during sync go into
the `pending` file and are synced on the next sync, which means it's safe to
delete the `syncing` file when sync completes.

Differential Revision: D7932299

fbshipit-source-id: bd86e836ded10e75790b87ca6734a29f068f3571