infinitepush: add test for excess file downloads during pullbackups
Summary:
If a user runs hg pullbackup while their repository doesn't have the
latest commits, it's possible that pullbackups will deliver some public commits
to the user in order to then deliver the necessary draft commits on top. Let's
add a test that verifies the public commits are not delivered with remotefilelog
data.

Reviewed By: quark-zju

Differential Revision: D7927966

fbshipit-source-id: 8661a291c8b3a67359c9975e130f23fec643307d