infinitepush: add test for excess treemanifest downloads
Summary:
This adds a test that verifies that if we pull infinitepush commits and
received some normal-public commits as part of that pull, we will only receive
the tree data for the infinitepush commits. The commit cloud team had some
concerns here, since this will be much more common in that future.

Reviewed By: singhsrb

Differential Revision: D7993229

fbshipit-source-id: c91752e0da86a81142503ed97ac104668d1af874