rebase: do not rely that every node in obsutil.allsuccessors present in the repo or changelog node map
Summary:
This is regulary not the case if Commit Cloud enabled, if a lot of work is done on one
machine 1, then cloud sync is done on another machine (machine 2), we sync draft not obsolete
heads and all history of obsmarkers, but there is no guarantee we download all
the obsoleted commits to the machine 2.

So, there can be a sequence of obsmarkers from commit A to commit B, when some commit in between hasn't been fetched to the machine 2.

Function obsutil.allsuccessors returns all successors (not only final successors) and it is intentionally do not check presence in the repo.

Reviewed By: DurhamG

Differential Revision: D8009823

fbshipit-source-id: f6b90011acaa518f71a2033eb2b885b27a7f06e7