asyncmemo: do not add key sizes to the total weight
Summary:
We had an asyncmemo bug when keys were, for example, Strings. If we insert
String::new("key"), and then try to remove String::new("key"), then these keys
may have different weights because Strings may have different capacity.

This is not the case for values though, because inserted and deleted value is
always the same object.

Let's not take key sizes into account at all. That means that the real asyncmemo memory usage will be
higher, however values are a great deal bigger than keys, so it sholdn't be an
issue.
Also we are planning to do it anyway, because we want to split
asyncmemo hashmap into in-flight futures and completed futures, and in-flight
futures won't take key sizes into account.

Reviewed By: jsgf

Differential Revision: D8040684

fbshipit-source-id: 332dbe5663182dd84ea291ca4b33b41995ac0166