treestate: avoid broad borrow on self in Node.visit
Summary:
Looking at the `path` variable, it is now:

  vec![KeyRef<'a>, KeyRef<'a>, KeyRef<'a>, ...]

where `'a` is bounded to `self`. That's okay but makes it impossible to
modify `self` between `path.push` and the end of the `visit` function.

For `Node<FileStateV2>`, `self` needs to be modified, because `visitor`
might change `file`'s state, and `self.aggregated_state` needs to be updated
accordingly.

Basically, the elements of `Vec` should match the call stack, and have
different lifetimes. A nested `visit()` should have a nested lifetime on
`Vec` elements, instead of relying on borrowing `self`.

  vec![KeyRef<'a>, KeyRef<'b>, KeyRef<'c>, ...]
  #    visit(path) {
  #                let name: KeyRef<'b>; // instead of <'a>.
  #                path.push(name)
  #                visit(path) {
  #                            let name: KeyRef<'c>;
  #                            path.push(name)
  #                            visit(path) { ... }
  #                            }
  #                }

The previous diff introduces a `VecStack` struct to work in this case. Let's
use it. This also make sure all `path.pop()` happens automatically even if
panic happens.

Reviewed By: markbt

Differential Revision: D7797457

fbshipit-source-id: e1723fac3dbd7d244b69f1fc46d90bef64a29a3f