treestate: rename watchman_clock to metadata
Summary:
The field is a string that contains information `TreeState` itself does not
care about. It's up to the upper layer to decide how to use it, and it does
not have to be watchman clock, and might contain other fields like hostname,
ignore matcher hash, etc. So let's rename it to clarify.

Differential Revision: D7886282

fbshipit-source-id: 739a85d7a710918e0b18a9b7fe0e31b366bab447