use better practice to pass authorization token in http header
Summary:
The querystring is encrypted with SSL

However, URLs can be stored in web server logs - typically the whole URL of each request is
stored in a server log. This means that any sensitive data in the URL (e.g. a
password) is being saved in clear text on the server

Reviewed By: farnz

Differential Revision: D8075914

fbshipit-source-id: 4eadd362755272414dadad069e7ba04db7a01ee8