update the Eden CLI to honor the EDEN_HG_BINARY environment variable
Summary:
When calling mercurial from inside the Eden CLI, use `$EDEN_HG_BINARY` as the
path to hg if this variable is set.  If it is not set, continue the current
behavior of using `hg` from the user's `$PATH`.  This is primarily used during
the `eden clone` command.

This makes sure the Eden integration tests use `hg` built from the local
repository rather than the version of `hg` installed on the system.  Most
locations in the integration tests were already doing so, but `eden clone` was
one place that still ended up using the system hg binary.

Reviewed By: wez

Differential Revision: D7839850

fbshipit-source-id: da801fad0767a111b3e3dfd393d82da8e2147e22