templatekw: add test demonstrating bad template expansion
Summary:
The template keyword '{succsandmarkers}', or any template keyword that
represents a list of dicts, does not expand well if used on its own.

This test demonstrates that it gets expanded to the string concatenation of the
object's keys, merged with the mapping dictionary's keys, which is very
meaningless.  Nobody seems to use this, except as a truthy value, but it seems
like bad behaviour.

The keys can come in any order, so the test just matches a long string of
characters.

In a future diff we will make this meaningful: the count of objects.

Reviewed By: mitrandir77

Differential Revision: D8161704

fbshipit-source-id: 78e1b86a64eb887603bd6e58682eb7567f06cbce