mononoke: add has_copyinfo
Summary:
Motivation for this change is the following:
We want to avoid doing point lookups to the db to fetch the history, and
instead select the history for a file or a directory with a single select. That
should be many times faster.
After we select the data from filenodes table, we need to fetch  copy info
data. To do that we need to either do lots of point lookups to the fixedcopyinfo (which
undermines the point of doing single select) or construct a giant select
statement. First option is inefficient, the second can introduce big code
complexity.

Instead let's add a single field that says whether we have fixed copy info or
not. Db size increase should be tiny.

For the data that we've already imported I'm planning to run a script that will
fill the has_copyinfo field correctly.

Reviewed By: jsgf

Differential Revision: D8164029

fbshipit-source-id: c91c99b065808a93a9b361914cf9b3822d78cb60