treemanifest: move diff calculation out of _getnewtree
Summary:
A future patch will improve the performance of convert flat manifests
to tree manifests by relying on the stored manifest delta instead of hydrating
the whole manifest. To do this, we need to change _getnewtree to not depend on
taking the entire flat manifests. So let's refactor it to take a diff instead,
and in a future patch we will optimize how we compute that diff.

Reviewed By: phillco

Differential Revision: D8121627

fbshipit-source-id: 41bbba6e98cd605ee185618b5f99883a9e0fd7cf