treemanifest: refactor new tree creation into separate parts
Summary:
This refactors the new tree creation logic from a single function to
two functions, one that computes the differences in (added, removed) format, and
another that accepts adds/removes and applies them to a copy of the parent tree.

This will help in a future diff where we want to unify code paths to use the
fast diff path.

Reviewed By: phillco

Differential Revision: D8122900

fbshipit-source-id: 82699dc4bbf69dd565f8564e71f8a0726659313e