mononoke: use get_all_filenodes
Summary:
In previous diffs we've added a function that does a bulk prefetch of
filenodes. Now we can use it in the getfiles request handling code.

Before constructing file history we do a bulk prefetch. Then we use
results of the bulk prefetch as a cache. It means that we first check
whether we have file history node in the prefetched data, and if not it will be
fetched again. It allows us to later plugin a memcache or some other kind of
cache.

repo.rs file is getting bigger and bigger. I'm planning to split it, however
only after I land this diff in order not to deal with merges

Reviewed By: farnz

Differential Revision: D8185701

fbshipit-source-id: 6ca37aeb029236db51d2b5a03cb7053f969cf47e