fix a bug in cdatapack fanout table calculation
Summary:
If all of the nodes in a datapack file start with the same byte value,
the cdatapack code computed the fanout table incorrectly.  If this byte value
was anything other than 0x00 it would only be able to find the last node in
the pack, and would not be able to find any other nodes.

This fixes the code to compute the fanout table correctly in this case.

Reviewed By: quark-zju

Differential Revision: D8131020

fbshipit-source-id: 84e49befc5776cff96831f6120194466d9c80b35