mononoke: refactor changed_entry_stream to not use recursive_entry_stream
Summary:
In the future diffs we'll add a trait called Pruner, which can be used to avoid
recursing into some paths. For example, to not recurse into deleted paths, not
recurse into the paths that ends on this string etc.

In this diff we just make a small refactoring to make adding Pruner easier.
The main point of the refactoring is to not call recursive_entry_stream
function.

Reviewed By: jsgf

Differential Revision: D8207269

fbshipit-source-id: 42f246a8a7aaf7a43c40fa05ebea876830391430