treemanifest: avoid ondemand conversions on the tree server
Summary:
We were encountering issues where readonly connections where failing
because the serve was returning a Permission denied error while fetching trees.
It turns out the ondemand tree generation was attempting to open a transaction
and create trees on the server.  Since the server should have all the trees
anyway, we shouldn't do this ondemand conversion.

We still need on demand conversions for infinitepush bundles, so we'll leave
that in.

Reviewed By: singhsrb

Differential Revision: D8226459

fbshipit-source-id: 2d05faefbe3ae92ccb81072e012673f29c9cf0d9