mononoke: add Pruner argument
Summary:
Pruner is an FnMut that is  called on every entry
that `changed_entry_stream` visits. If Pruner returns false, then
`changed_entry_stream` won't recurse into that.

The primary motivation for that is gettreepack request. We may get requests
with a few `mfnodes`. Currently we just merge a few `get_changed_entry`
streams. The problem is that some entries will be in both streams, and thus
will be generated twice. visited_pruner changes it.

Reviewed By: jsgf

Differential Revision: D8207271

fbshipit-source-id: 1896cb68edbca57651c4767d9326b9bbbab2c980