pushrebase: fix bundle pack file ordering
Summary:
Previously, when adding temporary pack files to the local data store,
we added them at the end. This causes problems with the recently added ondemand
generating datastores, since when we go down the line of stores we will hit the
generating store (which will create a tree from flat manifest, which is slow)
before we hit the bundle store that actually contains the data.

Let's reorder the stores so the bundle stores go at the front, since they are
likely to be used.

Reviewed By: phillco

Differential Revision: D8291415

fbshipit-source-id: 4f9799cf755b2b1f7fe8d8d1c041a3b3f2c7e463