hgrc: only enable remotenames.fastheaddiscovery for clients
Summary:
It could affect pushrebase behavior - pushrebase server might have a large
set of "missing" nodes, and take that as the commits to rebase. Then fail
with "cannot rebase public commits".

Maybe there is a better way to fix it so it only affects pull but not push.
But for now it seems safer to disable it for automation.

Reviewed By: DurhamG

Differential Revision: D8049229

fbshipit-source-id: 3612e2fb2c1c825e8974156c27b03966ee745c98